+++ CHANGELOG.markdown
@@ -2,6 +2,7 @@
 
 * Support for a "rebasefile" option to make file() calls with absolute files relative to another directory.
 * Fix an infinite loop with fqdn_rand
+* Recognize v5 hiera configuration format
 
 # v1.3.12.1 (2017/10/09)